moved static documentation files
