+++ frontend/hexacode-playground/public/js/theme.js
