+++ frontend/hexacode-playground/public/searchtools.js
